switch cowsay package, cow selection
use @wiisportsresorts/cowsay as cowsay provider

+++ src/commands/messages/cowsay.ts
@@ -1,4 +1,5 @@
-import { say } from 'cowsay2';
+import { say } from '@wiisportsresorts/cowsay';
+import * as cows from '@wiisportsresorts/cowsay/lib/cows';
 import { Message } from 'discord.js';
 import yargsParser from 'yargs-parser';
 
@@ -9,9 +10,12 @@ import { codeBlock, Embed } from '../../util';
 export class CommandCowsay implements Command {
   cmd = 'cowsay';
   yargs: yargsParser.Options = {
-    boolean: ['delete'],
+    boolean: ['delete', 'list'],
+    string: ['cow'],
     alias: {
       delete: 'd',
+      cow: 'f',
+      list: 'l',
     },
     default: {
       delete: false,
@@ -24,17 +28,29 @@ export class CommandCowsay implements Command {
   async execute(context: Context): Promise<void | Message> {
     const { msg, args } = context;
 
+    if (args.list)
+      return msg.channel.send(Embed.info('Cows', codeBlock(Object.keys(cows).join(', '))));
+
     if (args._.length == 0 || /^\s+$/.test(args._.join(' ')))
       return msg.channel.send(Embed.error('nothing to say'));
 
     args.delete && msg.deletable && msg.delete();
 
-    return msg.channel.send(
-      codeBlock(
-        say(args._.join(' '), {
-          W: 48,
-        })
-      )
-    );
+    const cow = Object.keys(cows).includes(args.cow)
+      ? cows[args.cow as keyof typeof import('@wiisportsresorts/cowsay/lib/cows')]
+      : undefined;
+
+    if (args.cow && !cow) return msg.channel.send(Embed.error('Unknown cow'));
+
+    const text = say(args._.join(' '), {
+      W: 48,
+      cow,
+    });
+
+    const messages = text
+      .match(/(.|\n){1,1990}\n/g)
+      ?.map(message => codeBlock(message)) as string[];
+
+    for (const text of messages) await msg.channel.send(text);
   }
 }
